requirements.txt
mysql-connector-python
Flask
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
openai
pytest
pytest-flask
